+++ RELEASE_NOTES.md
@@ -2,6 +2,7 @@
 
 ### TBD
 
+* Fixed an issue where `consumes` and `produces` values having a `charset` could break request/response validation *(Issue 295)*
 * Fixed an issue where with Swagger 2.0 documents, you cannot use `#validateModel` with an `Array` object *(Issue 303)*
 * Fixed an issue where array values were not validated against their validation constraints, on their format/type *(Issue 300)*
 * Fixed an issue with `collectionFormat=multi` not working right with singlar values *(Issue 313)*